Add intermediate saves (#227)
* Mostly fine- sometimes fails test_cluster.py?
