+++ enspara/cluster/hybrid.py
@@ -10,8 +10,19 @@ from . import kcenters
 from . import kmedoids
 from . import util
 
+#Circular import if entering this from cluster.py but need this if not.
+try:
+    from ..apps.cluster import write_assignments_and_distances_with_reassign, \
+    write_centers, write_centers_indices
+except:
+    pass
+
+from ..util.log import timed
+
 from ..exception import ImproperlyConfigured
 from .. import mpi
+from enspara.cluster.util import *
+
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +65,7 @@ class KHybrid(BaseEstimator, ClusterMixin, util.MolecularClusterMixin):
 
     def __init__(self, metric, n_clusters=None, cluster_radius=None,
                  kmedoids_updates=5, random_first_center=False,
-                 random_state=None, mpi_mode=None):
+                 random_state=None, mpi_mode=None, args=None, lengths=None):
 
         if n_clusters is None and cluster_radius is None:
             raise ImproperlyConfigured("Either n_clusters or cluster_radius "
@@ -68,8 +79,10 @@ class KHybrid(BaseEstimator, ClusterMixin, util.MolecularClusterMixin):
         self.metric = util._get_distance_method(metric)
         self.random_state = check_random_state(random_state)
         self.mpi_mode = mpi_mode if mpi_mode is not None else mpi.size() != 1
+        self.args = args
+        self.lengths = lengths
 
-    def fit(self, X, init_centers=None):
+    def fit(self, X, init_centers=None, args=None):
         """Takes trajectories, X, and performs KHybrid clustering.
         Optionally continues clustering from an initial set of cluster
         centers.
@@ -90,7 +103,8 @@ class KHybrid(BaseEstimator, ClusterMixin, util.MolecularClusterMixin):
             random_first_center=self.random_first_center,
             init_centers=init_centers,
             random_state=self.random_state,
-            mpi_mode=self.mpi_mode)
+            mpi_mode=self.mpi_mode, args=self.args,
+            lengths=self.lengths)
 
         self.runtime_ = time.perf_counter() - t0
 
@@ -100,7 +114,8 @@ class KHybrid(BaseEstimator, ClusterMixin, util.MolecularClusterMixin):
 def hybrid(
         X, distance_method, n_iters=5, n_clusters=np.inf,
         dist_cutoff=0, random_first_center=False,
-        init_centers=None, random_state=None, mpi_mode=False):
+        init_centers=None, random_state=None, mpi_mode=False,
+        args=None, lengths=None):
 
     distance_method = util._get_distance_method(distance_method)
 
@@ -113,10 +128,38 @@ def hybrid(
         result.center_indices, result.assignments, result.distances,
         result.centers)
 
+    print(args)
+    print(cluster_center_inds)
+
+    #TO DO - this is giving me a 1D array and expected is a 2D array.
+    if args != None and args.save_intermediates:
+
+        int_result = util.ClusterResult(
+            center_indices=cluster_center_inds,
+            assignments=assignments,
+            distances=distances,
+            centers=centers).partition(lengths)
+
+        int_indcs, int_assigs, int_dists, int_centers = int_result
+
+        print(int_indcs)
+        print(np.shape(int_assigs))
+        with timed("Wrote kcenters center indices in %.2f sec.", logger.info):
+            write_centers_indices(
+                args.center_indices,
+                [(t, f * args.subsample) for t, f in int_indcs],
+                intermediate_n=f'kcenters')
+
+        with timed("Wrote kcenters center structures in %.2f sec.", logger.info):
+            write_centers(int_result, args, intermediate_n=f'kcenters')
+
+        write_assignments_and_distances_with_reassign(int_result, args, 
+            intermediate_n=f'kcenters')
+
     if n_iters > 0:
         return kmedoids._kmedoids_iterations(
             X, distance_method, n_iters, cluster_center_inds, assignments,
-            distances)
+            distances, args=args, lengths=lengths)
     else:
         return util.ClusterResult(
             center_indices=cluster_center_inds,
